% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_profileid.R
\name{add_payment_type_column}
\alias{add_payment_type_column}
\title{add column for payment type}
\usage{
add_payment_type_column(dataset, payment_type)
}
\arguments{
\item{dataset}{(Ex. general_payments_2021)}

\item{payment}{type (ex. "general")}
}
\value{
an updated version of the dataset, with 2 new columns (year and payment_type) filled in with the relevant values
}
\description{
add column for payment type
}
