% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_profileid.R
\name{clean_physician_supplement}
\alias{clean_physician_supplement}
\title{clean physician supplement}
\usage{
clean_physician_supplement(dataset)
}
\arguments{
\item{dataset}{\itemize{
\item Covered Recipient Supplement File for All Program Years
available at: https://www.cms.gov/OpenPayments/Data/Dataset-Downloads
}}
}
\value{
cleaned version of this dataset - ie. filters out missing NPIs (~7k) & converts NPI to string
}
\description{
clean physician supplement
}
