% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_profileid.R
\name{clean_verified_npis}
\alias{clean_verified_npis}
\title{clean verified npis}
\usage{
clean_verified_npis(dataset)
}
\arguments{
\item{dataset}{where verified npis are stored in column "number" (ie. "verified_sample.rds")}
}
\value{
cleaned version of the dataset with missings & duplicates removed for ease of use with get_profile_ids()
}
\description{
clean verified npis
}
