% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_guideline_metadata.R
\name{get_guideline_metadata}
\alias{get_guideline_metadata}
\title{Get guideline metadata from Crossref
prereq: check crossref API user guide first & follow polite use rules}
\usage{
get_guideline_metadata(dataset)
}
\arguments{
\item{dataset}{guidelines dataframe that includes doi column named "doi"}
}
\value{
dataset of author names for each guideline doi in the argument dataset
}
\description{
Get guideline metadata from Crossref
prereq: check crossref API user guide first & follow polite use rules
}
