% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_profileid.R
\name{get_payment_records}
\alias{get_payment_records}
\title{get payment data}
\usage{
get_payment_records(dataset1, dataset2)
}
\arguments{
\item{dataset1}{(OP payments file of choice (general, research, ownership) for a given year)}

\item{dataset2}{(dataset returned by get_profile_ids())}
}
\value{
df of all payment records from dataset1 associated with a profile ID in dataset2
}
\description{
get payment data
}
