% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_profileid.R
\name{get_profile_ids}
\alias{get_profile_ids}
\title{get profile ids}
\usage{
get_profile_ids(dataset1, dataset2)
}
\arguments{
\item{dataset1}{(cleaned physician supplement - returned by clean_physician_supplement())}

\item{dataset2}{(cleaned verified npis - returned by clean_verified_npis())}
}
\value{
subset of unique profile ids in dataset1 where npis in dataset1 & dataset2 match
}
\description{
get profile ids
}
